requirements.txt
pyshorteners
psutil
pytest
qrcode
Pillow
speedtest-cli
pytest_xprocess